+++ sanity/schemaTypes/index.ts
@@ -1,5 +1,6 @@
 import { post } from './post'
 import { project } from './project'
 import { product } from './product'
+import { siteContent } from './siteContent'
 
-export const schemaTypes = [post, project, product]
+export const schemaTypes = [post, project, product, siteContent]